+++ Dockerfile
@@ -1,10 +1,10 @@
 # Build stage
-FROM golang:1.25.1-alpine AS builder
+FROM golang:1.26.5-alpine AS builder
 
 WORKDIR /go/src/app
 COPY . .
 
-RUN apk add git=2.49.1-r0
+RUN apk add git=2.54.0-r0
 
 RUN go build -ldflags="-X github.com/lidofinance/onchain-mon/internal/connectors/metrics.Commit=$(git rev-parse HEAD)" -o ./bin/feeder ./cmd/feeder
 RUN go build -ldflags="-X github.com/lidofinance/onchain-mon/internal/connectors/metrics.Commit=$(git rev-parse HEAD)" -o ./bin/forwarder ./cmd/forwarder
@@ -16,6 +16,5 @@ WORKDIR /app
 RUN apk add --no-cache ca-certificates
 
 COPY --from=builder /go/src/app/bin .
-COPY --from=builder /go/src/app/web /app/web/
 
 USER nobody
