+++ docker-compose.testnet.yaml
@@ -1,3 +1,17 @@
+# Testnet stack. Runs side by side with docker-compose.yml: every container
+# name, port, volume and project namespace is distinct, so both can be up at
+# once.
+#
+#   docker compose -f docker-compose.testnet.yaml --env-file .env.testnet up -d
+#
+# Needs .env.testnet (copy sample.testnet.env) and notification.testnet.yaml
+# with testnet channels. Defaults target Hoodi:
+#   BLOCK_TOPIC="blocks.hoodi.l1"
+#   JSON_RPC_URL=https://hoodi.drpc.org
+#   BLOCK_EXPLORER=hoodi.etherscan.io
+
+name: onchain-mon-testnet
+
 x-logging: &default-logging
   options:
     max-size: "50m"
@@ -8,31 +22,33 @@ services:
     extends:
       file: docker-compose.base.yaml
       service: redis
-    container_name: redis
+    container_name: redis-testnet
     ports:
-      - "6379:6379"
+      - "6380:6379"
+    volumes:
+      - redis-testnet-data:/data
 
   nats:
     extends:
       file: docker-compose.base.yaml
       service: nats
-    container_name: nats
+    container_name: nats-testnet
     ports:
-      - "4222:4222"
-      - "8222:8222"
+      - "4223:4222"
+      - "8223:8222"
     volumes:
-      - ./infra/nats/jetstream:/data/jetstream
+      - ./infra/nats/jetstream-testnet:/data/jetstream
       - ./infra/nats/nats.conf:/etc/nats/nats.conf
 
   forwarder:
     image: lidofinance/onchain-mon:stable
-    container_name: forwarder
+    container_name: forwarder-testnet
     build: ./
     restart: always
     command:
       - ./forwarder
     env_file:
-      - .env
+      - .env.testnet
     environment:
       - READ_ENV_FROM_SHELL=true
       - ENV=${ENV}
@@ -45,7 +61,7 @@ services:
       - REDIS_DB=${REDIS_DB}
       - QUORUM_SIZE=${QUORUM_SIZE}
     ports:
-      - "8081:8080"
+      - "8083:8080"
     depends_on:
       - redis
       - nats
@@ -54,16 +70,17 @@ services:
       # /etc/forwarder/notification.yaml otherwise, so mount both.
       - ./notification.yaml:/app/notification.yaml
       - ./notification.yaml:/etc/forwarder/notification.yaml
+    logging: *default-logging
 
   feeder:
     image: lidofinance/onchain-mon:stable
-    container_name: feeder
+    container_name: feeder-testnet
     build: ./
     restart: always
     command:
       - ./feeder
     env_file:
-      - .env
+      - .env.testnet
     environment:
       - READ_ENV_FROM_SHELL=true
       - ENV=${ENV}
@@ -75,7 +92,11 @@ services:
       - REDIS_ADDRESS=redis:6379
       - BLOCK_TOPIC=${BLOCK_TOPIC}
     ports:
-      - "8082:8080"
+      - "8084:8080"
     depends_on:
       - redis
       - nats
+    logging: *default-logging
+
+volumes:
+  redis-testnet-data:
