+++ infra/nats/nats.conf
@@ -9,13 +9,17 @@ port: 4222
 http_port: 8222
 
 # JetStream enabled
-jetstream: true
+# jetstream: true
 
 # JetStream storage directory
-store_dir: "/data/jetstream"
+jetstream {
+  store_dir: /data
+  max_memory_store: 128mb
+  max_file_store: 1GB
+}
 
 # Maximum message size
-max_payload: 3MB
+max_payload: 8MB
 
 # Logging
 # Log level: "debug", "info", "warn", "error", "fatal"
