+++ sample.env
@@ -1,23 +1,29 @@
+# Mainnet template. Copy to .env and fill in.
+# Every key here is read by internal/env/env.go — nothing else is.
+
+# When true, config comes from the shell instead of this file (docker-compose
+# sets it to true and passes the values through `environment:`).
 READ_ENV_FROM_SHELL=false
+
+# Must be UNIQUE per forwarder instance — quorum breaks if two instances share it.
 SOURCE=local
 
 ENV=local
-APP_NAME=finding_forwarder
+APP_NAME=onchain_mon
 PORT=8080
 LOG_FORMAT=simple
 LOG_LEVEL=debug
 
+NATS_DEFAULT_URL="http://localhost:4222"
 BLOCK_TOPIC="blocks.mainnet.l1"
 
-SENTRY_DSN="SENTRY_DSN"
-
-NATS_DEFAULT_URL="http://service-forta-nats:4222"
-REDIS_ADDRESS="redis:6379"
+REDIS_ADDRESS="localhost:6379"
 REDIS_DB=0
+
+# How many instances must see a finding before it is sent (prod: 2 of 3).
 QUORUM_SIZE=1
 
 JSON_RPC_URL=https://eth.drpc.org
-
 BLOCK_EXPLORER=etherscan.io
 
-STETH_NATS_PUBLISH_TOPIC="findings.protocol.steth"
+SENTRY_DSN=""
