+++ sample.testnet.env
@@ -1,0 +1,27 @@
+# Hoodi testnet template. Copy to .env.testnet and fill in.
+# Same keys as sample.env — only the network-specific values differ.
+
+READ_ENV_FROM_SHELL=false
+
+# Must be UNIQUE per forwarder instance — quorum breaks if two instances share it.
+SOURCE=local-hoodi
+
+ENV=local
+APP_NAME=onchain_mon_hoodi
+PORT=8080
+LOG_FORMAT=simple
+LOG_LEVEL=debug
+
+NATS_DEFAULT_URL="http://localhost:4223"
+BLOCK_TOPIC="blocks.hoodi.l1"
+
+# Separate Redis DB keeps testnet quorum keys away from mainnet ones.
+REDIS_ADDRESS="localhost:6380"
+REDIS_DB=1
+
+QUORUM_SIZE=1
+
+JSON_RPC_URL=https://hoodi.drpc.org
+BLOCK_EXPLORER=hoodi.etherscan.io
+
+SENTRY_DSN=""
